zerowaste/requirements.txt
asgiref==3.3.1
astor==0.8.1
attrs==21.2.0
boto3==1.17.78
botocore==1.20.78
bson==0.5.10
bytecode==0.12.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
Django==3.0.10
django-allauth==0.44.0
django-dotenv==1.4.2
django-extensions==3.1.2
django-filter==2.4.0
django-mysql==3.11.1
django-mysql-geventpool==0.2.5
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gevent==21.1.2
graphviz==0.16
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jellyfish==0.8.2
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
networkx==2.5.1
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pygraphviz==1.7
PyJWT==1.7.1
pynguin==0.8.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rich==10.2.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
simple-parsing==0.0.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.6.0
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.4
zope.event==4.5.0
zope.interface==5.4.0
